+++ src/xrdtiffoperations/math/DataMasking.java
@@ -46,8 +46,8 @@ public class DataMasking {
         return result;
     }
 
-    private static short maskValue(short _value, short _max, short _min){
-        short value;
+    private static int maskValue(int _value, int _max, int _min){
+        int value;
 
         value = _value;
         if (value < _min){
